+++ src/main/java/com/masil/domain/address/repository/EmdAddressRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface EmdAddressRepository extends JpaRepository<EmdAddress, Long> {
+public interface EmdAddressRepository extends JpaRepository<EmdAddress, Integer> {
     List<EmdAddress> findByEmdNameContains(String search);
 }
