+++ src/main/java/com/masil/domain/address/repository/SggAddressRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SggAddressRepository extends JpaRepository<SggAddress , Long> {
+public interface SggAddressRepository extends JpaRepository<SggAddress , Integer> {
 
     List<SggAddress> findBySggNameContains(String search);
